Refactor move functions
